Refactor ApiClient to unwrap results and throw errors.
The previous behavior required the caller of ApiClient to check
both the `result` key in the returned object for the string
`success`, and to catch errors in a `catch` block. This change cleans
up the code by allowing the consumer of ApiClient to only handle thrown
errors. In addition, the consumer receives an unwrapped result
(such as api_key for a fetchApiKey call) instead of a wrapped object.