Check for URL validity in apiFetch
Calling 'fetch' with invalid URL in iOS correctly throws an
exception. On Android though it throws inside Java code and
leads to a crash.

Ideally we will catch invalid URLs even earler, but this adds an
important guard clause and additional info in the exception,
ensuring it is thrown within our JS code.