docs: Explain how to promote a beta release to prod.
Like the other stages of the process, this is a lot more complicated
for iOS than for Android / Google Play.

Contrary to what I briefly thought yesterday based on some hurried
reading on the web, though, it isn't so perverse that it requires
re-uploading the artifact -- instead, much like on Google Play, the
same artifact already out to TestFlight can be sent to the App Store.
Which means the whole process (assuming it passes Apple's review and
doesn't require a revised build...) can indeed be done entirely
through the web.