webview: Group closely related scroll-handling code together.
The `isMessageNode` helper is used only by `getStartAndEndNodes`,
which in turn is used only in `sendScrollMessage` and the initializer
for `prevNodes`; so put them right by that code.

Similarly, the one purpose of `handleScrollEvent` is to be installed
as an event listener, so put the `addEventListener` call right after
the definition.