webview: Rename a helper and flip its sense: showHideElement.
The verb "toggle" means turning something on if it's off, and off if
it's on -- flipping it from whatever state it's currently in.  The
meaning of this function isn't that but rather to set the thing to
some definite state the caller asks for, regardless of what state it
might already be in.  So, adjust the name to match the meaning.

(The DOM method `toggle` is named to reflect what it does with only
one argument; it's unfortunate that with two arguments it means
something at odds with its name.  But we can stop the confusion here.)

Also flip the sense of the function.  Every call site was computing
a value for whether to *show* the element, then negating it before
passing here; so this lets us remove a bunch of double negatives,
making the call sites more straightforward to read.