webview: Use the new HTML-escaping template tag everywhere.
This flips things around so that *not* escaping is marked, while
escaping is boring and unmarked, as it should be.  The better to make
non-escaped interpolation stand out, we use the new tag in absolutely
every template literal in `src/webview/html/` that produces HTML.

We even include templates with no interpolation at all.  They might
grow interpolation in the future, they may serve as models for
contributors adding other templates which do have interpolation, and
the consistency helps in confirming we didn't miss any: try running
  $ git grep -P '(?<!template)`(?!;)' src/webview/html/
and you find only closing backticks and a few in the tag's own tests.