message: Mark fetching event when narrowing from notification.
Most fetching of messages goes through `fetchMessages`, which
dispatches a "message fetch start" event first so that we can show
loading indicators instead of emptiness.

Narrowing from a notification is a special case where we load at
startup the messages required for the message narrow, and we do so
in `fetchEssentialInitialData`, which bypasses `fetchMessages`.
This wasn't setting correctly that "fetching" state, so we were
displaying "No messages" while actually loading new messages.

Fixes #2714.

[greg: expanded commit message]