flow: Update reselect stubs from `flow-typed`.
The `Selector` definition has been split into `InputSelector` and
`OutputSelector`, so we have to adjust a bit of code accordingly.
Also, instead of `any` we pass `void` as the second parameter, as
this selector does not depend on `props` passed to it.

We also get to remove a now-unnecessary $FlowFixMe.