flow: Push down Action type fixmes to specific locations.
In the previous few commits, we eliminated the remaining `any` types
that made up Action, and left a fixme marker to temporarily preserve
the status quo in hiding type errors that a proper type-checking of
uses of Action would find.

Remove that, so that Action now must be one of the specified proper
types.  This causes Flow to error in a few places.  All are because
our type `Dispatch` is too strict about the return type of a thunk
action creator, requiring it to return only void or Promise<void>.

[greg: split from parent, added comments]