webview: Increase line-height within message subheader.
A line-height of 1 was cutting off the bottom of certain characters
such as 'p' and 'g'. Remove the line-height declarations within
the .username and .timestamp rule-sets so that line-height for the
text selected by those classes is inherited from the document body
(which is set to 1.4 at the time of this commit).

Fixes #2696.