flow: Remove disused `Actions` type.
Before we migrated to standard Redux usage of `dispatch` and `connect`
with #2628, we used this type routinely to describe the `actions` prop
of many components.  Following that, and after the last straggling use
of it was removed just now in b0d0097c9, it no longer has any
references at all.  Remove it.