conversations: Get types on `getRecentConversations`.
This function is kind of complicated, and Flow wasn't figuring out the
types so it wasn't able to help us understand.  Fix that.

In fact this turns up a type confusion in the code: we're looking up
strings in an object that's really meant to have numbers as keys.
Because for plain JS objects the keys (formally: the names of
properties) are always strings, and anything else is converted with
`toString`, this works at runtime -- but it sure is confusing when
trying to understand how the code is supposed to work.  Explain in a
fixme comment.