persist: Add explicit list of store keys we don't persist.
Add 'discardKeys' that lists all the rest of the state keys that
we intentionally do not store or cache. This has no effect but is
to make it more explicit. We can also add some tests that check if
all keys we do have are in one of the three structures.