android: Let release and debug builds coexist on the same device.
This changes the app ID for a debug build, so the Android system sees
it as a separate app and lets it be installed at the same time as the
release version.

There is one workflow annoyance: the `react-native run-android`
command is not very smart.  The app gets built and installed just
fine, because those are handled within Gradle; but then `run-android`
has to start the app, and it uses the wrong app ID, `com.zulipmobile`.
This will start the release build if you have it on the device
(potentially extremely confusing, as the edits you just made to the
code don't seem to show up...), or just give an error if you don't.

A workaround is to use this command instead:

  $ react-native run-android --appIdSuffix=debug

Fortunately we don't need to document that, because we'll follow up
with a hack to make it unnecessary instead.

Fixes #2705.

[greg: rewrote commit message]