pms: Don't list a group PM conversation with missing data.
This is consistent with how we do it for 1:1 PM conversations.

Previously in `GroupPmConversationItem` we would try to retrieve all
recipients from the `usersByEmail` data. If not found, we substituted
them with a `NULL_USER`. This resulted in empty names, which looked
obviously broken; and the items wouldn't work for narrowing to the
actual conversation, either.

Now, we first search for all users. If any of the users are not found
we do not render the component at all.