users: Remove unused function fetchUsers().
The only reference to fetchUsers() was removed with 8362d617e in 2016
(the function was named sendGetUsers() at that time).  fetchUsers was
also (since b1bd9c7b7 in 2017-10) the only place where INIT_USERS got
dispatched, so we can remove that action as well.

Currently, we load an initial list of all users when initializing an
account, and update it on EVENT_USER_ADD events.