users: Remove nonexistent is_active property from User.
The server used to supply an `is_active` property for each user.
This is no longer the case; since the server's 305fe6106 (in
1.8.0), the information is now expressed by the split between
`realm_active_users` and `realm_non_active_users`.

We are already using `realm_non_active_users`, but we hadn't
cleaned out the old, obsolete code yet.  Clean that out.