android: Fool rn-cli into starting the right app.
Our purpose here is to make just plain `react-native run-android`
do the right thing again, rather than have to tell everyone to
pass `--appIdSuffix=debug` or something equivalently nonobvious
now that we set the application ID of debug builds to
`com.zulipmobile.debug`.

The reason this doesn't work in the first place is that the RN CLI
is using a silly heuristic to figure out the ID of our app, which
gets it wrong here.  See the code:
  https://github.com/facebook/react-native/blob/46ffb1062/local-cli/runAndroid/runAndroid.js#L111

The piece of this code that does the main work is that line with
`package=` in a comment at the top of the manifest source file.
This exploits the same silly heuristic's limitations to tell it
the right answer.  Because it's in an XML comment, none of the
real build process notices or is affected.

But then fixing that causes it to make an assumption in turn about
the name of our main activity, and try to pass an intent filter
like `com.zulipmobile.debug/com.zulipmobile.debug.MainActivity`
to `am start -n`.

So we add a sort of alias to make that name work.

[greg: squashed fixup commits in; revised commit message;
 added comments in manifest.]