api: Add all back-end options to `registerForEvents`.
Add a link to the API call documentation on zulipchat.com and add
support for all properties the back-end supports, while making them
optional by setting their values to the defaults.

Accept all parameters as an object instead of what would need to be
a very long list of nameless function arguments.  Use the new
`objectToParams` function to prepare this object for the back-end.