scripts: Update calls to `npm` to use `yarn`.
This better aligns our code with our docs, so that we use just `yarn`
everywhere rather than `npm`, and hopefully make things a bit less
confusing for new contributors who aren't already Node developers
familiar with the precise relationship between those two commands.

As a nice bonus, the output of commands like `yarn test` loses some
gratuitous noise.

While we're adjusting these commands, take the opportunity to make
`yarn test` print a line marking the end of lint and the beginning of
the Jest tests, by replacing just `jest` with `yarn run jest`.

Note that for scripts, we don't use the convenience syntax where `run`
is omitted; while that's fine for interactive use, it seems dodgy to
use in code because the command could be shadowed by a future
subcommand of `yarn` itself; and in fact the upstream `yarn run` doc
warns against doing so.