emoji: Knock out for now our broken realm-emoji support.
Realm emoji don't currently work in any useful way in the
autocompletion we provide in composing a message; if you try typing
one, it just won't show up in the list.  For more details and context,
see #2846.

To simplify reasoning about this code as we go and extend it to
provide an emoji picker for reacting with an emoji, cut out a few
bits of code that are wrong and especially confusing:

* getFilteredEmojiList gets passed a data structure which is indexed
  by the emoji's numeric identifiers; but treats those identifiers as
  the emoji's names, so they end up in its output (and can in fact be
  seen in the autocomplete list, if you start typing digits.)

* The way `EmojiRow` handles realm emoji doesn't make much sense: it
  takes a complete `RealmEmojiType` object with all the details, but
  just treats it like a boolean, and the `RealmEmoji` child goes and
  gets things from the state again.

This doesn't affect the area where realm emoji work correctly today,
which is in rendering messages that contain them.