msg actions: Cut never-used `onlyIf` logic for header sheet.
This was copied from the similar function below it, but it's not
actually used.  Simplify things by cutting it out.

I think probably it'd be cleaner to replace the `onlyIf` logic
entirely by this style, but that's for another time.

Also cut the punning use of `skip` for `onPress` here, allowing us to
simplify the type annotation to reflect what the handlers really do.