docs/build-run: Make an editing pass over JDK version issues.
We now have both a "Java too old" and "Java too new" issue; make an
editing pass over their text, and add a bit more context.

I hadn't actually known the story with short-term versus LTS Java
releases until I read up on this just now!

Also, while it's reasonable for a given Gradle version not to work out
of the box on a JDK version that didn't exist when it was released,
it's clearly wrong that it blows up with this cryptic error message.
But since Gradle 4.7 (also the first to support Java 10), there is
hope that it won't do that: https://github.com/gradle/gradle/pull/4759