rehydration: Allow `payload` to be undefined.
In redux-persist v5, the `payload` of the rehydrate action can be
null/undefined, at least on the first boot-up.  Add `undefined` to the
type, and adjust our one reducer that didn't already allow for this.
This doesn't affect the behavior of our code for redux-persist v4.

Background: redux-persist v5 handles the initial rehydration
differently.  Previously, the initial rehydration was the job of the
internal function adhocRehydrate().  adhocRehydrate() initialized the
state without accessing the underlying permanent storage.  With v5,
that logic got simplified -- like all other rehydrations, the initial
rehydration initializes the store with what's stored in the permanent
storage.  This means that the payload broadcast with the rehydrate
action is now actually null/undefined on the first boot-up.