jest: Fix Jest config to work with Babel 7.
Instead of getting our Jest preprocessor from the `babel-jest`
package, use the one supplied by React Native.  This will keep Jest
working when it starts using Babel 7, with the RN v0.57 upgrade.

See discussion in facebook/react-native#19859 .  This is seen there as
a workaround, but it seems like it might simply be the right solution.
In any case it's the most understandable solution we've been able to find.